+++ client/package.json
@@ -10,6 +10,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "client": "file:",
     "framer-motion": "^11.11.9",
     "jwt-decode": "^4.0.0",
     "react": "^18.3.1",
