+++ server/package.json
@@ -22,6 +22,7 @@
     "multer": "^1.4.5-lts.1",
     "multer-storage-cloudinary": "^4.0.0",
     "nodemailer": "^6.9.15",
-    "nodemon": "^3.1.6"
+    "nodemon": "^3.1.6",
+    "server": "file:"
   }
 }
