+++ server/routes/user.route.js
@@ -3,12 +3,8 @@ const authMiddleware = require("../middlewares/authMiddleware");
 const {
   getUser,
   getUserByRole,
-  createUser,
-  loginUser,
   updateActive,
-  verifyOTP,
   updateUser,
-  resendOtp,
 } = require("../controllers/user.controller");
 const upload = require("../middlewares/uploadImage");
 const router = express.Router();
@@ -17,11 +13,7 @@ const router = express.Router();
 router.get("/", authMiddleware(["admin"]), getUser);
 router.get("/:role", authMiddleware(["admin"]), getUserByRole);
 router.post("/activeAccount", authMiddleware(["admin"]), updateActive);
-//Đăng ký, Đăng nhập cho Tenant và Landlord
-router.post("/", upload.single("image"), createUser);
-router.post("/login", loginUser);
-router.post("/verifyOTP", verifyOTP);
-router.post("/resendOtp", resendOtp);
+
 
 //Chỉnh sửa thông tin (Tenant và Landlord)
 router.post(
